rename all test_processing_*.py to test_processor_*.py ans fix duplicate test processor paligemma
