+++ tests/models/fuyu/test_processor_fuyu.py
