+++ tests/models/idefics2/test_processor_idefics2.py
