+++ tests/models/idefics3/test_processor_idefics3.py
