+++ tests/models/musicgen/test_processor_musicgen.py
