+++ tests/models/qwen2_vl/test_processor_qwen2_vl.py
